Fix analyzer warnings
