+++ src/Sextant.Avalonia/NavigationView.cs
@@ -4,28 +4,24 @@
 // See the LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reactive;
 using System.Reactive.Concurrency;
 using System.Reactive.Linq;
+
 using Avalonia;
-using Avalonia.Animation;
 using Avalonia.Controls;
 using Avalonia.Media;
-using Avalonia.ReactiveUI;
 using Avalonia.Styling;
-using DynamicData.Aggregation;
-using DynamicData.Binding;
+
 using ReactiveUI;
-using Sextant;
 
 namespace Sextant.Avalonia
 {
     /// <summary>
     /// The <see cref="IView"/> implementation for Avalonia.
     /// </summary>
-    public sealed class NavigationView : ContentControl, IStyleable, IView
+    public sealed partial class NavigationView : ContentControl, IStyleable, IView
     {
         private readonly Navigation _modalNavigation = new Navigation();
         private readonly Navigation _pageNavigation = new Navigation();
@@ -56,7 +52,7 @@ namespace Sextant.Avalonia
                         {
                             _modalNavigation.Control
                         },
-                        [!Panel.BackgroundProperty] = 
+                        [!Panel.BackgroundProperty] =
                             _modalNavigation
                                 .CountChanged
                                 .Select(count => count > 0 ? new SolidColorBrush(Colors.Transparent) : null)
@@ -143,88 +139,6 @@ namespace Sextant.Avalonia
 
             view.ViewModel = viewModel;
             return view;
-        }
-
-        /// <summary>
-        /// This class represents a single navigation layer.
-        /// </summary>
-        private class Navigation
-        {
-            private readonly ObservableCollection<IViewFor> _navigationStack = new ObservableCollection<IViewFor>();
-            private readonly IPageTransition _transition;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="Navigation"/> helper.
-            /// </summary>
-            public Navigation()
-            {
-                var navigationStackObservable = _navigationStack
-                    .ToObservableChangeSet()
-                    .Publish()
-                    .RefCount();
-
-                navigationStackObservable
-                    .Select(_ => _navigationStack.LastOrDefault())
-                    .Subscribe(page => Control.Content = page);
-
-                CountChanged = navigationStackObservable.Count().AsObservable();
-                _transition = Control.PageTransition;
-            }
-
-            /// <summary>
-            /// Toggles the animations.
-            /// </summary>
-            /// <param name="enable">Returns true if we are enabling the animations.</param>
-            public void ToggleAnimations(bool enable) => Control.PageTransition = enable ? _transition : null;
-
-            /// <summary>
-            /// Gets a bool indicating whether the control is visible.
-            /// </summary>
-            public bool IsVisible => _navigationStack.Count > 1;
-
-            /// <summary>
-            /// Publishes new values when page count changes.
-            /// </summary>
-            public IObservable<int> CountChanged { get; }
-
-            /// <summary>
-            /// The control responsible for rendering the current view.
-            /// </summary>
-            public TransitioningContentControl Control { get; } = new TransitioningContentControl();
-
-            /// <summary>
-            /// Adds a <see cref="IViewFor"/> to the navigation stack.
-            /// </summary>
-            /// <param name="view">The view to add to the navigation stack.</param>
-            /// <param name="resetStack">Defines if we should reset the navigation stack.</param>
-            public void Push(IViewFor view, bool resetStack = false)
-            {
-                if (resetStack)
-                {
-                    _navigationStack.Clear();
-                }
-
-                _navigationStack.Add(view);
-            }
-
-            /// <summary>
-            /// Removes the last <see cref="IViewFor"/> from the navigation stack.
-            /// </summary>
-            public void Pop()
-            {
-                var indexToRemove = _navigationStack.Count - 1;
-                _navigationStack.RemoveAt(indexToRemove);
-            }
-
-            /// <summary>
-            /// Removes all pages except the first one <see cref="IViewFor"/> from the navigation stack.
-            /// </summary>
-            public void PopToRoot()
-            {
-                var view = _navigationStack[0];
-                _navigationStack.Clear();
-                _navigationStack.Add(view);
-            }
         }
     }
 }
